Monument HW (Very Late)
Worked tied me up and ended up having this left in the dust a bit. A lot of errors, but you can sort of see I was well on the way to getting to the proper layout, just didn't have enough attention towards it because other things sort of piled on.